Revert initialize_file_logger function
