+++ convert2rhel/initialize.py
@@ -49,31 +49,6 @@ def initialize_logger():
     return logger_module.setup_logger_handler()
 
 
-def initialize_file_logging(log_name, log_dir):
-    """
-    Archive existing file logs and setup all logging handlers that require
-    root, like FileHandlers.
-
-    This function should be called after
-    :func:`~convert2rhel.main.initialize_logger`.
-
-    .. warning::
-        Setting log_dir underneath a world-writable directory (including
-        letting it be user settable) is insecure.  We will need to write
-        some checks for all calls to `os.makedirs()` if we allow changing
-        log_dir.
-
-    :param str log_name: Name of the logfile to archive and log to
-    :param str log_dir: Directory where logfiles are stored
-    """
-    try:
-        logger_module.archive_old_logger_files(log_name, log_dir)
-    except (IOError, OSError) as e:
-        loggerinst.warning("Unable to archive previous log: %s" % e)
-
-    logger_module.add_file_handler(log_name, log_dir)
-
-
 def set_locale():
     """
     Set the C locale, also known as the POSIX locale, for the main process as well as the child processes.
@@ -117,10 +92,6 @@ def run():
 
     # initialize logging
     initialize_logger()
-
-    # since we now have root, we can add the FileLogging
-    # and also archive previous logs
-    initialize_file_logging("convert2rhel.log", logger_module.LOG_DIR)
 
     from convert2rhel import main
