+++ convert2rhel/unit_tests/main_test.py
@@ -29,22 +29,9 @@ from convert2rhel import backup
 six.add_move(six.MovedModule("mock", "mock", "unittest.mock"))
 from six.moves import mock
 
-from convert2rhel import (
-    actions,
-    applock,
-    checks,
-    exceptions,
-    grub,
-    main,
-    pkghandler,
-    pkgmanager,
-    redhatrelease,
-    repo,
-    subscription,
-    toolopts,
-    unit_tests,
-    utils,
-)
+from convert2rhel import actions, applock, checks, exceptions, grub
+from convert2rhel import logger as logger_module
+from convert2rhel import main, pkghandler, pkgmanager, redhatrelease, repo, subscription, toolopts, unit_tests, utils
 from convert2rhel.actions import report
 from convert2rhel.breadcrumbs import breadcrumbs
 from convert2rhel.systeminfo import system_info
@@ -53,6 +40,7 @@ from convert2rhel.unit_tests import (
     CLIMocked,
     CollectEarlyDataMocked,
     FinishCollectionMocked,
+    InitializeFileLoggingMocked,
     MainLockedMocked,
     PrintDataCollectionMocked,
     PrintSystemInformationMocked,
@@ -132,6 +120,35 @@ class TestRollbackChanges:
             main.rollback_changes()
 
 
+@pytest.mark.parametrize(("exception_type", "exception"), ((IOError, True), (OSError, True), (None, False)))
+def test_initialize_file_logging(exception_type, exception, monkeypatch, caplog):
+    add_file_handler_mock = mock.Mock()
+    archive_old_logger_files_mock = mock.Mock()
+
+    if exception:
+        archive_old_logger_files_mock.side_effect = exception_type
+
+    monkeypatch.setattr(
+        logger_module,
+        "add_file_handler",
+        value=add_file_handler_mock,
+    )
+    monkeypatch.setattr(
+        logger_module,
+        "archive_old_logger_files",
+        value=archive_old_logger_files_mock,
+    )
+
+    main.initialize_file_logging("convert2rhel.log", "/tmp")
+
+    if exception:
+        assert caplog.records[-1].levelname == "WARNING"
+        assert "Unable to archive previous log:" in caplog.records[-1].message
+
+    add_file_handler_mock.assert_called_once()
+    archive_old_logger_files_mock.assert_called_once()
+
+
 class TestShowEula:
     eula_dir = os.path.realpath(os.path.join(os.path.dirname(__file__), "..", "data", "version-independent"))
 
@@ -185,6 +202,7 @@ def test_help_exit(monkeypatch, tmp_path):
     monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
     monkeypatch.setattr(sys, "argv", ["convert2rhel", "--help"])
     monkeypatch.setattr(utils, "require_root", RequireRootMocked())
+    monkeypatch.setattr(main, "initialize_file_logging", InitializeFileLoggingMocked())
     monkeypatch.setattr(main, "main_locked", MainLockedMocked())
 
     with pytest.raises(SystemExit):
@@ -195,6 +213,7 @@ def test_help_exit(monkeypatch, tmp_path):
 
 def test_main(monkeypatch, tmp_path):
     require_root_mock = mock.Mock()
+    initialize_file_logging_mock = mock.Mock()
     toolopts_cli_mock = mock.Mock()
     show_eula_mock = mock.Mock()
     print_data_collection_mock = mock.Mock()
@@ -220,6 +239,7 @@ def test_main(monkeypatch, tmp_path):
 
     monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
     monkeypatch.setattr(utils, "require_root", require_root_mock)
+    monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
     monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
     monkeypatch.setattr(main, "show_eula", show_eula_mock)
     monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -245,6 +265,7 @@ def test_main(monkeypatch, tmp_path):
 
     assert main.main() == 0
     assert require_root_mock.call_count == 1
+    assert initialize_file_logging_mock.call_count == 1
     assert toolopts_cli_mock.call_count == 1
     assert show_eula_mock.call_count == 1
     assert print_data_collection_mock.call_count == 1
@@ -270,6 +291,7 @@ def test_main(monkeypatch, tmp_path):
 class TestRollbackFromMain:
     def test_main_rollback_post_cli_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
+        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock(side_effect=Exception)
 
@@ -278,12 +300,14 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
+        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "finish_collection", finish_collection_mock)
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
+        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert finish_collection_mock.call_count == 1
@@ -292,6 +316,7 @@ class TestRollbackFromMain:
     def test_main_traceback_in_clear_versionlock(self, caplog, monkeypatch, tmp_path):
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", RequireRootMocked())
+        monkeypatch.setattr(main, "initialize_file_logging", InitializeFileLoggingMocked())
         monkeypatch.setattr(toolopts, "CLI", CLIMocked())
         monkeypatch.setattr(main, "show_eula", ShowEulaMocked())
         monkeypatch.setattr(breadcrumbs, "print_data_collection", PrintDataCollectionMocked())
@@ -331,6 +356,7 @@ class TestRollbackFromMain:
 
     def test_main_traceback_before_action_completion(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
+        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -349,6 +375,7 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
+        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -365,6 +392,7 @@ class TestRollbackFromMain:
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
+        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -385,6 +413,7 @@ class TestRollbackFromMain:
 
     def test_main_rollback_pre_ponr_changes_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
+        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -405,6 +434,7 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
+        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -423,6 +453,7 @@ class TestRollbackFromMain:
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
+        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -442,6 +473,7 @@ class TestRollbackFromMain:
 
     def test_main_rollback_analyze_exit_phase(self, global_tool_opts, monkeypatch, tmp_path):
         require_root_mock = mock.Mock()
+        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -461,6 +493,7 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
+        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -480,6 +513,7 @@ class TestRollbackFromMain:
 
         assert main.main() == 0
         assert require_root_mock.call_count == 1
+        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
@@ -496,6 +530,7 @@ class TestRollbackFromMain:
 
     def test_main_rollback_post_ponr_changes_phase(self, monkeypatch, caplog, tmp_path):
         require_root_mock = mock.Mock()
+        initialize_file_logging_mock = mock.Mock()
         toolopts_cli_mock = mock.Mock()
         show_eula_mock = mock.Mock()
         print_data_collection_mock = mock.Mock()
@@ -518,6 +553,7 @@ class TestRollbackFromMain:
 
         monkeypatch.setattr(applock, "_DEFAULT_LOCK_DIR", str(tmp_path))
         monkeypatch.setattr(utils, "require_root", require_root_mock)
+        monkeypatch.setattr(main, "initialize_file_logging", initialize_file_logging_mock)
         monkeypatch.setattr(toolopts, "CLI", toolopts_cli_mock)
         monkeypatch.setattr(main, "show_eula", show_eula_mock)
         monkeypatch.setattr(breadcrumbs, "print_data_collection", print_data_collection_mock)
@@ -538,6 +574,7 @@ class TestRollbackFromMain:
 
         assert main.main() == 1
         assert require_root_mock.call_count == 1
+        assert initialize_file_logging_mock.call_count == 1
         assert toolopts_cli_mock.call_count == 1
         assert show_eula_mock.call_count == 1
         assert print_data_collection_mock.call_count == 1
